requirements.txt
cufflinks
pandas
streamlit
numpy
biopython
requests